pyproject.toml
[project]
name = "wedding-res"
version = "0.1.0"
description = ""
authors = [
    {name = "Mwapsam",email = "mwapsam@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "django (>=5.2.4,<6.0.0)",
    "django-extensions (>=4.1,<5.0)",
    "qrcode (>=8.2,<9.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "django-allauth[socialaccount] (>=65.10.0,<66.0.0)",
    "django-redis (>=6.0.0,<7.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "django-storages (>=1.14.6,<2.0.0)",
    "whitenoise (>=6.9.0,<7.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "boto3 (>=1.39.9,<2.0.0)",
    "celery (>=5.5.3,<6.0.0)",
    "redis (>=6.2.0,<7.0.0)",
    "django-celery-results (>=2.6.0,<3.0.0)",
    "django-celery-beat (>=2.8.1,<3.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "django-htmx (>=1.23.2,<2.0.0)",
    "fido2 (>=2.0.0,<3.0.0)",
    "phonenumbers (>=9.0.10,<10.0.0)"
]

[tool.poetry]
packages = [{include = "wedding_res", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
